third_party/WebKit/LayoutTests/fast/html/imports/import-style-link-block.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<link id="importLink" rel="import" href="resources/style-link-block-child.html">
<script>
shouldBeNonNull("importLink.import");
</script>
