third_party/WebKit/LayoutTests/fast/html/imports/import-style-link-block-expected.txt

PASS childStyleLink.sheet.title is "Script Blocker"
PASS importLink.import is non-null.
PASS successfullyParsed is true

TEST COMPLETE
